requirements.txt
langchain
openai
streamlit
streamlit-chat
langchain-openai
python-dotenv
PyPDF2
tiktoken
faiss-cpu
faiss-gpu